# PERF: avoid object conversion in fillna(method=pad|backfill) for masked arrays