--- a/pandas/_libs/algos.pyx
+++ b/pandas/_libs/algos.pyx
@@ -625,6 +625,38 @@ def pad_inplace(algos_t[:] values, const uint8_t[:] mask, limit=None):
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
+def pad_inplace_masked(algos_t[:] values, uint8_t[:] mask, limit=None):
+    cdef:
+        Py_ssize_t i, N
+        algos_t val
+        uint8_t prev_mask
+        int lim, fill_count = 0
+
+    N = len(values)
+
+    # GH#2778
+    if N == 0:
+        return
+
+    lim = validate_limit(N, limit)
+
+    val = values[0]
+    prev_mask = mask[0]
+    for i in range(1, N):
+        if mask[i]:
+            if fill_count >= lim:
+                continue
+            fill_count += 1
+            values[i] = val
+            mask[i] = prev_mask
+        else:
+            fill_count = 0
+            val = values[i]
+            prev_mask = mask[i]
+
+
+@cython.boundscheck(False)
+@cython.wraparound(False)
 def pad_2d_inplace(algos_t[:, :] values, const uint8_t[:, :] mask, limit=None):
     cdef:
         Py_ssize_t i, j, N, K