--- a/pandas/core/arrays/base.py
+++ b/pandas/core/arrays/base.py
@@ -708,7 +708,7 @@ class ExtensionArray:
         if mask.any():
             if method is not None:
                 func = get_fill_func(method)
-                new_values = func(self.astype(object), limit=limit, mask=mask)
+                new_values, _ = func(self.astype(object), limit=limit, mask=mask)
                 new_values = self._from_sequence(new_values, dtype=self.dtype)
             else:
                 # fill with value