--- a/pandas/tests/extension/test_interval.py
+++ b/pandas/tests/extension/test_interval.py
@@ -133,6 +133,10 @@ class TestMissing(BaseInterval, base.BaseMissingTests):
         pass
 
     @unsupported_fill
+    def test_fillna_no_op_returns_copy(self):
+        pass
+
+    @unsupported_fill
     def test_fillna_series(self):
         pass