```
[ 75.00%] ··· frame_methods.Fillna.time_frame_fillna                                                                         ok
[ 75.00%] ··· ========= ======== ============ ============= ========== ============= =============
              --                                               dtype                              
              ------------------ -----------------------------------------------------------------
               inplace   method    float64       float32      object       Int64        Float64   
              ========= ======== ============ ============= ========== ============= =============
                 True     pad     3.44±0.3ms    2.64±0.2ms   307±10ms    3.97±0.2ms    3.72±0.1ms 
                 True    bfill    3.23±0.2ms    2.89±0.2ms   275±8ms     2.34±0.1ms   2.45±0.09ms 
                False     pad     3.62±0.1ms   2.06±0.01ms   181±3ms    4.23±0.07ms   4.16±0.07ms 
                False    bfill    3.59±0.2ms   2.06±0.02ms   186±6ms    4.35±0.09ms   4.18±0.06ms 
              ========= ======== ============ ============= ========== ============= =============
```

```
       before           after         ratio
     [2e5c28fc]       [d3ae0cf1]
     <master>         <BaseMaskedArray.fillna>
-     2.28±0.06ms      2.06±0.02ms     0.91  frame_methods.Fillna.time_frame_fillna(False, 'bfill', 'float32')
-      15.5±0.2ms      2.45±0.09ms     0.16  frame_methods.Fillna.time_frame_fillna(True, 'bfill', 'Float64')
-      16.5±0.2ms       2.34±0.1ms     0.14  frame_methods.Fillna.time_frame_fillna(True, 'bfill', 'Int64')
-      85.2±0.9ms      4.18±0.06ms     0.05  frame_methods.Fillna.time_frame_fillna(False, 'bfill', 'Float64')
-      86.6±0.6ms      4.16±0.07ms     0.05  frame_methods.Fillna.time_frame_fillna(False, 'pad', 'Float64')
-      90.9±0.8ms      4.35±0.09ms     0.05  frame_methods.Fillna.time_frame_fillna(False, 'bfill', 'Int64')
-      92.4±0.4ms      4.23±0.07ms     0.05  frame_methods.Fillna.time_frame_fillna(False, 'pad', 'Int64')
-        85.3±1ms       3.72±0.1ms     0.04  frame_methods.Fillna.time_frame_fillna(True, 'pad', 'Float64')
-        92.5±2ms       3.97±0.2ms     0.04  frame_methods.Fillna.time_frame_fillna(True, 'pad', 'Int64')

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```