@jreback 2021-02-22
looks good. pls add a whatsnew note. you had a comment about multiple copies ; can try to handle here or make a new issue. ping when ready.